fix: Added GSLIB executable
Added MPS algorithms from GSLIB.
