--- wgan/wgan.py
@@ -1,169 +1,0 @@
-import albumentations as A
-from torchvision.transforms.transforms import GaussianBlur
-import os
-import argparse
-import numpy as np
-import time
-
-import torchvision.transforms as transforms
-from torchvision.utils import save_image
-from torchvision.datasets import ImageFolder
-
-from torch.autograd import Variable
-
-import torch.nn as nn
-import torch
-
-os.makedirs("images", exist_ok=True)
-os.makedirs("batch_images", exist_ok=True)
-
-parser = argparse.ArgumentParser()
-parser.add_argument("--n_epochs", type=int, default=2000, help="number of epochs of training")
-parser.add_argument("--batch_size", type=int, default=16, help="size of the batches")
-parser.add_argument("--lr", type=float, default=0.0002, help="adam: learning rate")
-parser.add_argument("--b1", type=float, default=0.5, help="adam: decay of first order momentum of gradient")
-parser.add_argument("--b2", type=float, default=0.999, help="adam: decay of first order momentum of gradient")
-parser.add_argument("--n_cpu", type=int, default=1, help="number of cpu threads to use during batch generation")
-parser.add_argument("--latent_dim", type=int, default=100, help="dimensionality of the latent space")
-parser.add_argument("--img_size", type=int, default=128, help="size of each image dimension")
-parser.add_argument("--channels", type=int, default=1, help="number of image channels")
-parser.add_argument("--n_critic", type=int, default=5, help="number of training steps for discriminator per iter")
-parser.add_argument("--clip_value", type=float, default=0.01, help="lower and upper clip value for disc. weights")
-parser.add_argument("--sample_interval", type=int, default=50, help="interval betwen image samples")
-opt = parser.parse_args()
-print(opt)
-
-
-img_shape = (opt.channels, opt.img_size, opt.img_size)
-
-cuda = True if torch.cuda.is_available() else False
-
-
-class Generator(nn.Module):
-    def __init__(self):
-        super(Generator, self).__init__()
-
-        def block(in_feat, out_feat, normalize=True):
-            layers = [nn.Linear(in_feat, out_feat)]
-            if normalize:
-                layers.append(nn.BatchNorm1d(out_feat, 0.8))
-            layers.append(nn.LeakyReLU(0.2, inplace=True))
-            return layers            
-
-        self.model = nn.Sequential(
-            *block(opt.latent_dim, 128, normalize=False),
-            *block(128, 256),
-            *block(256, 512),
-            *block(512, 1024),
-            nn.Linear(1024, int(np.prod(img_shape))),
-            nn.Tanh()
-        )
-
-    def forward(self, z):
-        img = self.model(z)
-        img = img.view(img.shape[0], *img_shape)
-        return img
-
-
-class Discriminator(nn.Module):
-    def __init__(self):
-        super(Discriminator, self).__init__()
-
-        self.model = nn.Sequential(
-            nn.Linear(int(np.prod(img_shape)), 512),
-            nn.LeakyReLU(0.2, inplace=True),
-            nn.Linear(512, 256),
-            nn.LeakyReLU(0.2, inplace=True),
-            nn.Linear(256, 1),
-        )
-
-    def forward(self, img):
-        img_flat = img.view(img.shape[0], -1)
-        validity = self.model(img_flat)
-        return validity
-
-
-# Initialize generator and discriminator
-generator = Generator()
-discriminator = Discriminator()
-
-if cuda:
-    generator.cuda()
-    discriminator.cuda()
-
-# Configure data loader
-transform = transforms.Compose([
-    transforms.GaussianBlur(kernel_size=(3, 3)),
-    transforms.Grayscale(num_output_channels=1),
-    transforms.Resize((opt.img_size, opt.img_size)), transforms.ToTensor(),
-    transforms.RandomErasing(p=0.3),
-    transforms.Normalize([0.5], [0.5])])
-
-dataset = ImageFolder(r"..\data\\", transform=transform)
-dataloader = torch.utils.data.DataLoader(dataset, batch_size=opt.batch_size, shuffle=True)
-
-# Optimizers
-optimizer_G = torch.optim.RMSprop(generator.parameters(), lr=opt.lr)
-optimizer_D = torch.optim.RMSprop(discriminator.parameters(), lr=opt.lr)
-
-Tensor = torch.cuda.FloatTensor if cuda else torch.FloatTensor
-
-# ----------
-#  Training
-# ----------
-
-batches_done = 0
-for epoch in range(opt.n_epochs):
-    for i, (imgs, _) in enumerate(dataloader):
-        # Configure input
-        real_imgs = Variable(imgs.type(Tensor))
-
-        # ---------------------
-        #  Train Discriminator
-        # ---------------------
-
-        optimizer_D.zero_grad()
-
-        # Sample noise as generator input
-        z = Variable(Tensor(np.random.normal(0, 1, (imgs.shape[0], opt.latent_dim))))
-        #print(f"Z shape {z.shape}")
-        # Generate a batch of images
-        fake_imgs = generator(z).detach()
-        #print(f"Fake images shape {fake_imgs.shape}")
-        #print(f"Real images shape {real_imgs.shape}")
-        # Adversarial loss
-        loss_D = -torch.mean(discriminator(real_imgs)) + torch.mean(discriminator(fake_imgs))
-
-        loss_D.backward()
-        optimizer_D.step()
-
-        # Clip weights of discriminator
-        for p in discriminator.parameters():
-            p.data.clamp_(-opt.clip_value, opt.clip_value)
-
-        # Train the generator every n_critic iterations
-        if i % opt.n_critic == 0:
-
-            # -----------------
-            #  Train Generator
-            # -----------------
-            optimizer_G.zero_grad()
-
-            # Generate a batch of images
-            gen_imgs = generator(z)
-            # Adversarial loss
-            loss_G = -torch.mean(discriminator(gen_imgs))
-
-            loss_G.backward()
-            optimizer_G.step()
-
-            print(
-                "[Epoch %d/%d] [Batch %d/%d] [D loss: %f] [G loss: %f]"
-                % (epoch, opt.n_epochs, batches_done % len(dataloader), len(dataloader), loss_D.item(), loss_G.item())
-            )
-
-        if batches_done % opt.sample_interval == 0:
-            save_image(gen_imgs.data[:25], "batch_images/%d.png" % batches_done, nrow=5, normalize=True)
-            for idx, im in enumerate(gen_imgs):
-                save_image(im.data, f"images/{idx}_{batches_done}_{epoch}.png")
-        batches_done += 1
